+++ src/main/java/com/woowa/woowakit/domain/order/domain/OrderStatus.java
@@ -4,5 +4,5 @@ import lombok.Getter;
 
 @Getter
 public enum OrderStatus {
-	ORDERED, PLACED, CANCELED, PAYED,
+	ORDERED, CANCELED, PAYED,
 }
