+++ src/main/java/com/woowa/woowakit/domain/payment/domain/Payment.java
@@ -33,31 +33,25 @@ public class Payment extends BaseEntity {
 	@Convert(converter = MoneyConverter.class)
 	private Money totalPrice;
 
-	private String uuid;
-
 	@Builder
 	private Payment(
 		final String paymentKey,
 		final Money totalPrice,
-		final String uuid,
 		final Long orderId
 	) {
 		this.paymentKey = paymentKey;
 		this.totalPrice = totalPrice;
-		this.uuid = uuid;
 		this.orderId = orderId;
 	}
 
 	public static Payment of(
 		final String paymentKey,
 		final Money totalPrice,
-		final String uuid,
 		final Long orderId
 	) {
 		return Payment.builder()
 			.paymentKey(paymentKey)
 			.totalPrice(totalPrice)
-			.uuid(uuid)
 			.orderId(orderId)
 			.build();
 	}
