+++ src/main/java/com/woowa/woowakit/infra/payment/toss/ThreadSleepClient.java
@@ -1,9 +1,11 @@
 package com.woowa.woowakit.infra.payment.toss;
 
-import com.woowa.woowakit.domain.model.Money;
-import com.woowa.woowakit.domain.payment.domain.PaymentClient;
 import java.time.Duration;
 import java.util.Random;
+
+import com.woowa.woowakit.domain.model.Money;
+import com.woowa.woowakit.domain.order.domain.PaymentClient;
+
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Mono;
@@ -25,7 +27,7 @@ public class ThreadSleepClient implements PaymentClient {
 		final Money totalPrice
 	) {
 		long latancyMs =
-			(long) ((LATENCY_MIN + STANDARD_DEVIATION * new Random().nextGaussian()) * 1000);
+			(long)((LATENCY_MIN + STANDARD_DEVIATION * new Random().nextGaussian()) * 1000);
 		log.info("결제 요청 반환에 {} ms 가 수행됩니다. paymentKey: {}", latancyMs, paymentKey);
 
 		return
