+++ src/main/resources/schema.sql
@@ -75,7 +75,6 @@ create table if not exists payments
     order_id bigint not null,
     payment_key varchar(255) not null,
     total_price bigint not null,
-    uuid varchar(255) not null,
     created_at datetime not null,
     updated_at datetime not null,
     primary key (id)
